finished backend logic for twitter scraping/sentiment analysis
Will scrape tweets from the twitter API.
Check if the user is in our DB, if not add them, localise their location and attempt to identify their state.
Checks if the tweet is already in our db.
Run sentiment analysis on the tweet.
Saves the tweet and the analysis to the db.